refactor: Schema Refactor
