+++ src/schemas/mod.rs
@@ -1,2 +1,8 @@
+pub(crate) mod common;
+pub(crate) mod configuration_management;
+pub(crate) mod fonts;
 pub(crate) mod help_schema;
-pub(crate) mod sdb_schema;
+pub(crate) mod os_settings;
+pub(crate) mod shell_configuration;
+pub(crate) mod state_file;
+pub(crate) mod tools;
